frontend redesigned, realtime delete works, realtime add and search left

+++ jPhotosBackend/src/main/java/com/jPhotos/jPhotosBackend/repository/ImageRepository.java
@@ -9,9 +9,10 @@ import org.springframework.stereotype.Repository;
 import com.jPhotos.jPhotosBackend.model.ImageTable;
 
 @Repository
-public interface ImageRepository extends JpaRepository<ImageTable, Integer> {
+public interface ImageRepository extends JpaRepository<ImageTable, Long> {
 //	Optional<ImageTable> findByName(String name);
 //	List<ImageTable> findAll();
 	List<ImageTable> findAll();
     List<ImageTable> findByNameLike(String name);
+    
 }
